+++ lib/mongoid/history/trackable.rb
@@ -52,7 +52,7 @@ module Mongoid
         end
 
         def track_history?
-          Mongoid::History.enabled? && Thread.current[track_history_flag] != false
+          Mongoid::History.enabled? && Mongoid::History.store[track_history_flag] != false
         end
 
         def dynamic_enabled?
@@ -60,10 +60,10 @@ module Mongoid
         end
 
         def disable_tracking(&_block)
-          Thread.current[track_history_flag] = false
+          Mongoid::History.store[track_history_flag] = false
           yield
         ensure
-          Thread.current[track_history_flag] = true
+          Mongoid::History.store[track_history_flag] = true
         end
 
         def track_history_flag